Compatibility matrix change
After reviewing https://docs.djangoproject.com/en/2.0/faq/install/#what-python-version-can-i-use-with-django I’ve adjusted what python 3 needs to work with: the outstanding bug seems to be a python3 / django 1.8 issue, which is an unlikley combination.

We only need to support 1.10 and 1.11 for public consumption, but I’m aware of some Ramlwrap / Django 1.9 users.